NHDR-250 refactor: GPT&CLOVA 연동 코드 리팩토링
