+++ src/main/java/org/scoula/config/RootConfig.java
@@ -60,7 +60,8 @@ import lombok.extern.log4j.Log4j2;
 	"org.scoula.statistics.service",
 	"org.scoula.statistics.scheduler",
 	"org.scoula.View.Event.Service",
-	"org.scoula.news.service"
+	"org.scoula.news.service",
+	"org.scoula.question.service"
 })
 @Log4j2
 @EnableTransactionManagement
